requirements.txt
tensorflow_gpu==1.7.0
numpy==1.14.2
opencv_python==3.4.0.12
